✅ Fix: Dashboard fonctionnel - Backend/Frontend opérationnels
- Correction variables d'environnement API_URL (IP publique)
- Configuration CORS pour frontend (91.98.131.218:3001)
- Fix imports database.py (services.database)
- Configuration métriques Prometheus avec registre personnalisé
- Correction colonnes SQL (payout->actual_profit, profit_loss->actual_profit)
- Renommage champs stats (nb_bets->total_bets, etc.)
- Fix schémas Pydantic (Decimal->float pour BankrollSummary)
- Correction hook useUpdateBet (PUT->PATCH)

Status: Dashboard affiche 20 opportunités réelles, stats fonctionnelles
Endpoints opérationnels: /stats/bankroll, /opportunities, /bets
Erreurs 404 restantes: endpoints futurs non critiques (compare-agents, tips, etc.)

+++ backend/api/models/schemas.py
@@ -143,13 +143,13 @@ class Opportunity(OpportunityResponse):
 
 class BankrollSummary(BaseModel):
     """Résumé du bankroll"""
-    current_bankroll: Decimal
+    current_bankroll: float
     total_bets: int
     won_bets: int
     lost_bets: int
     pending_bets: int
-    total_staked: Decimal
-    total_profit: Decimal
+    total_staked: float
+    total_profit: float
     roi: float
     win_rate: float
